修复幻灵铁砧会被磁铁吸引的问题，修改幻灵铁砧实体碰撞箱

+++ src/main/java/dev/dubhe/anvilcraft/block/MagnetBlock.java
@@ -4,6 +4,7 @@ import dev.dubhe.anvilcraft.AnvilCraft;
 import dev.dubhe.anvilcraft.api.hammer.IHammerRemovable;
 import dev.dubhe.anvilcraft.entity.AnimateAscendingBlockEntity;
 
+import dev.dubhe.anvilcraft.init.ModBlocks;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
@@ -22,7 +23,6 @@ import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.phys.AABB;
 
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -97,7 +97,7 @@ public class MagnetBlock extends Block implements IHammerRemovable {
             currentPos = currentPos.below();
             BlockState state1 = level.getBlockState(currentPos);
 
-            if (state1.is(BlockTags.ANVIL)) {
+            if (state1.is(BlockTags.ANVIL) && !state1.is(ModBlocks.SPECTRAL_ANVIL)) {
                 level.destroyBlock(magnetPos.below(), true);
                 level.setBlockAndUpdate(magnetPos.below(), state1);
                 level.setBlockAndUpdate(currentPos, Blocks.AIR.defaultBlockState());
@@ -109,7 +109,7 @@ public class MagnetBlock extends Block implements IHammerRemovable {
                 level.getEntitiesOfClass(FallingBlockEntity.class, new AABB(currentPos));
             for (FallingBlockEntity entity : entities) {
                 BlockState state2 = entity.getBlockState();
-                if (state2.is(BlockTags.ANVIL)) {
+                if (state2.is(BlockTags.ANVIL) && !state2.is(ModBlocks.SPECTRAL_ANVIL)) {
                     level.destroyBlock(magnetPos.below(), true);
                     level.setBlockAndUpdate(magnetPos.below(), state2);
                     entity.remove(Entity.RemovalReason.DISCARDED);
